# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Warriors Gym Membership API
  version: 1.0.0
  description: |
    REST API behind the Warriors Gym admin web app, admin mobile app and member mobile app.

    **Authentication.** Sign in at `/auth/sign-in/{userType}`, then send the returned token as
    `Authorization: Bearer <token>`. Admin routes need an admin token and member routes a member
    token; a token for one role is rejected on the other.

    **Errors.** Every error is `{ "error": "message" }` with a meaningful status
    (400 validation, 401 auth, 403 forbidden, 404 missing, 409 conflict, 413 too large, 429 rate limited).

    **Demo.** When the server runs in demo mode, sample data is seeded and the two published demo
    logins cannot be changed. Data resets nightly.
servers:
  - url: http://localhost:5000
    description: Local development
...
paths:
...
  /auth/sign-up/{userType}:
    post:
      tags: [Auth]
      summary: Register a member (or an admin with an invite code)
      description: |
        Members can register freely. Admin registration is disabled unless the server has
        `ADMIN_INVITE_CODE` set, and then requires it in the `x-invite-code` header.
        Fields outside the documented set are ignored.
      parameters:
        - $ref: "#/components/parameters/userType"
        - in: header
          name: x-invite-code
          description: Required only for `admin`
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignUp" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties: { user: { type: string, example: user created successfuly } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /auth/sign-in/{userType}:
    post:
      tags: [Auth]
      summary: Sign in and receive a JWT
      parameters:
        - $ref: "#/components/parameters/userType"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone_number, password]
              properties:
                phone_number: { type: string }
                password: { type: string }
                language: { type: string, description: Saved as the user's language }
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string, description: JWT carrying id, phone_number and role }
                  userCheck:
                    oneOf:
                      - $ref: "#/components/schemas/Member"
                      - $ref: "#/components/schemas/Admin"
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /auth/sign-up/{phoneNumber}:
    get:
      tags: [Auth]
      summary: Check whether a member phone number is registered
      parameters:
        - in: path
          name: phoneNumber
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                type: object
                properties: { user: { type: string, example: phone number exists } }
        "404": { $ref: "#/components/responses/NotFound" }
  /auth/self:
    get:
      tags: [Auth]
      summary: Check that a token is still valid
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties: { token: { type: string, example: correct } }
        "401":
          description: Missing, malformed, invalid or expired token
          content:
            application/json:
              schema:
                type: object
                properties: { message: { type: string } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    userType:
      in: path
      name: userType
      required: true
      schema: { type: string, enum: [member, admin] }
...
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Forbidden:
      description: Not allowed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties: { error: { type: string } }
...
    SignUp:
      type: object
      required: [full_name, phone_number, password, language]
      properties:
        full_name: { type: string }
        phone_number: { type: string }
        password: { type: string, minLength: 6 }
        language: { type: string }
        gender: { type: string, description: Member only }
        height: { type: integer, description: Member only }
        weight: { type: integer, description: Member only }
        age: { type: integer, description: Member only }
        registration_Date: { type: string, description: Member only }
        admin_level: { type: string, description: Admin only }
    Member:
      type: object
      properties:
        id: { type: string, example: "00042", description: Zero-padded }
        full_name: { type: string }
        phone_number: { type: string }
        gender: { type: string }
        height: { type: integer }
        weight: { type: integer }
        age: { type: integer }
        activity_status: { type: string, enum: [Active, Payment Due, Inactive] }
        language: { type: string }
        image: { type: string, nullable: true }
        registration_Date: { type: string }
    Admin:
      type: object
      properties:
        id: { type: string, format: uuid }
        full_name: { type: string }
        phone_number: { type: string }
        admin_level: { type: string }
        language: { type: string }
        image: { type: string, nullable: true }
...